+++ cert-scanner/reporters/metrics_test.go
@@ -23,7 +23,7 @@ type MetricsReporterTests struct {
 	suite.Suite
 	sut                      *MetricsReporter
 	testResult               *TestCertResult
-	receivedDuration         float64
+	receivedDuration         float32
 	receivedTimingLabels     map[string]string
 	receivedValidationLabels map[string]string
 }
@@ -49,7 +49,7 @@ func (t *MetricsReporterTests) TestReportsSuccessfulResultAsMetric() {
 	result.Duration = time.Duration(123)
 	t.sut.Report(context.Background(), result)
 
-	t.Equal(float64(1.23e+08), t.receivedDuration)
+	t.Equal(float32(1.23e+08), t.receivedDuration)
 	t.Equal(map[string]string{"failed": "false", "foo": "bar", "source": "SomePod-acdf-bdfe", "source_type": "kubernetes"}, t.receivedTimingLabels)
 	t.Equal(map[string]string(nil), t.receivedValidationLabels)
 }
@@ -60,12 +60,12 @@ func (t *MetricsReporterTests) TestReportsFailureResultAsMetric() {
 	result.Duration = time.Duration(123)
 	t.sut.Report(context.Background(), result)
 
-	t.Equal(float64(1.23e+08), t.receivedDuration)
+	t.Equal(float32(1.23e+08), t.receivedDuration)
 	t.Equal(map[string]string{"failed": "true", "foo": "bar", "source": "SomePod-acdf-bdfe", "source_type": "kubernetes"}, t.receivedTimingLabels)
 	t.Equal(map[string]string{"failed": "true", "foo": "bar", "source": "SomePod-acdf-bdfe", "source_type": "kubernetes", "type": "some-error", "common_name": "somehost", "id": "4d2"}, t.receivedValidationLabels)
 }
 
-func (t *MetricsReporterTests) mockTimingMetric(duration float64, labels map[string]string) {
+func (t *MetricsReporterTests) mockTimingMetric(duration float32, labels map[string]string) {
 	t.receivedDuration = duration
 	t.receivedTimingLabels = labels
 }
